Split out tokenizer and test it explicit.
